\name{combinations}

\alias{combinations}

\title{Function to calculate all possible combinations of a vector of integers}

\description{bla bla bla}

\usage{combinations(id)}

\arguments{
  \item{id}{vector of integers} }

\details{space-time correlation}

\value{two vectors representing all possible combinations}

\references{Koenig et al. (1998)}

\author{Daniel Doktor}

\keyword{utilities}

\examples{\dontrun{combinations(id)}}
