\name{equaly}

\alias{equaly}

\title{Function to determine years where both stations observed}

\description{Coming shortly ...}

\usage{equaly(obs1,year1,obs2,year2)}

\arguments{
  \item{obs1}{date of budburst as julian day (integer)} 
  \item{year1}{year of observation (integer)}
  \item{obs2}{date of budburst as julian day (integer)}
  \item{year2}{year of observation (integer)} }

\details{not necessary}

\value{two vectors of containing budburst datess}

\author{Daniel Doktor}

\keyword{utilities}

\examples{\dontrun{equaly(obs1,year1,obs2,year2)}}
